Last Version

+++ Forces/Area.cs
@@ -4,7 +4,7 @@ namespace Forces
 {
     public class Area : Materia
     {
-        public Area(string name, int density, Color color) : base(name, density) => Color = color;
+        public Area(string name, double density, Color color) : base(name, density) => Color = color;
 
         public Color Color { get; set; }
     }
